MNT: Move FileHandler back to mixins.file_io for now
